tsfresh_exploration/JNJ/optimization/requirements.txt
pandas
numpy
matplotlib
talib-binary
alpaca-py
python-dotenv
pytz
